feat(kepad): implement new keyboard features and refactor UI components - Introduced a "fast number" feature in the keyboard extension for Android, enhancing user interaction. - Refactored gesture detection logic in `KepadModyil` to improve performance and reliability. - Updated `AngolImeService` to manage lifecycle and state more effectively with ViewModel and SavedStateRegistry. - Added new `kepad` module with Jetpack Compose support for a modern UI. - Enhanced hexagonal widget calculations for better rendering and interaction. - Updated Gradle configurations to include necessary dependencies and support for Kotlin Multiplatform.

+++ kepad/src/commonMain/kotlin/wedjets/AwdirRenqWedjet.kt
@@ -54,7 +54,7 @@ fun AwdirRenqWedjet(
                         secondaryLabel = if (isLetterMode && longPressLabel.isNotEmpty()) longPressLabel else null,
                         backgroundColor = hexColor,
                         textColor = KepadKonfeg.getComplementaryColor(hexColor),
-                        size = with(density) { geometry.hexWidth.toDp() },
+                        size = with(density) { geometry.hexWidth.toFloat().toDp() },
                         rotationAngle = geometry.rotationAngle.toFloat(),
                         onTap = { onHexKeyPress(tapLabel, false, null) },
                         onLongPress = if (isLetterMode && longPressLabel.isNotEmpty()) {
